requirements.txt
pandas
matplotlib
seaborn
playwright
openpyxl
python-dotenv
